useChartsOptionsからフックス以外の関数をlibに移しました。
